+++ src/main/java/com/pintoss/auth/api/order/dto/OrderDetailResponse.java
@@ -2,9 +2,7 @@ package com.pintoss.auth.api.order.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.pintoss.auth.core.order.application.dto.OrderDetail;
-import com.pintoss.auth.core.order.application.dto.OrderItemDetail;
-import com.pintoss.auth.core.order.domain.OrderStatus;
-import com.pintoss.auth.core.payment.application.PaymentMethodType;
+
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.Data;
