+++ src/main/java/com/pintoss/auth/client/billgate/ApprovalCreditResponseParser.java
@@ -5,8 +5,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.galaxia.api.merchant.Message;
 import com.pintoss.auth.common.exception.ErrorCode;
 import com.pintoss.auth.common.exception.server.InternalServerException;
-import com.pintoss.auth.core.payment.application.PaymentApprovalResponse;
-import com.pintoss.auth.core.payment.application.PaymentMethodType;
+import com.pintoss.auth.core.payment.application.dto.PaymentApprovalResponse;
+import com.pintoss.auth.core.payment.domain.PaymentMethodType;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.Map;
