+++ src/main/java/com/pintoss/auth/client/billgate/ApprovalResponseParserFactory.java
@@ -1,7 +1,7 @@
 package com.pintoss.auth.client.billgate;
 
 import com.pintoss.auth.common.exception.client.BadRequestException;
-import com.pintoss.auth.core.payment.application.PaymentMethodType;
+import com.pintoss.auth.core.payment.domain.PaymentMethodType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
