+++ src/main/java/com/pintoss/auth/common/event/PaymentCompletedEvent.java
@@ -1,6 +1,6 @@
 package com.pintoss.auth.common.event;
 
-import com.pintoss.auth.core.payment.application.PaymentMethodType;
+import com.pintoss.auth.core.payment.domain.PaymentMethodType;
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
