--- src/main/java/com/pintoss/auth/core/payment/application/PaymentApprovalService.java
@@ -1,6 +1,0 @@
-package com.pintoss.auth.core.payment.application;
-
-public interface PaymentApprovalService {
-
-    PaymentApprovalResponse approval(PaymentApprovalRequest request);
-}
