+++ src/main/java/com/pintoss/auth/core/payment/application/flow/event/PaymentEventPublisher.java
@@ -1,4 +1,4 @@
-package com.pintoss.auth.core.payment.application;
+package com.pintoss.auth.core.payment.application.flow.event;
 
 import com.pintoss.auth.common.event.PaymentCompletedEvent;
