+++ src/main/java/com/pintoss/auth/core/payment/application/flow/external/PaymentApprover.java
@@ -1,0 +1,9 @@
+package com.pintoss.auth.core.payment.application.flow.external;
+
+import com.pintoss.auth.core.payment.application.dto.PaymentApprovalRequest;
+import com.pintoss.auth.core.payment.application.dto.PaymentApprovalResponse;
+
+public interface PaymentApprover {
+
+    PaymentApprovalResponse approval(PaymentApprovalRequest request);
+}
