+++ src/main/java/com/pintoss/auth/event/SpringPaymentEventPublisher.java
@@ -1,7 +1,7 @@
 package com.pintoss.auth.event;
 
 import com.pintoss.auth.common.event.PaymentCompletedEvent;
-import com.pintoss.auth.core.payment.application.PaymentEventPublisher;
+import com.pintoss.auth.core.payment.application.flow.event.PaymentEventPublisher;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Component;
